[K6.0] User Manager: Move messages loads a site without options #8552
