+++ src/admin/src/Controller/UserController.php
@@ -26,6 +26,7 @@ use Joomla\Utilities\ArrayHelper;
 use Kunena\Forum\Libraries\Access\KunenaAccess;
 use Kunena\Forum\Libraries\Factory\KunenaFactory;
 use Kunena\Forum\Libraries\Forum\Category\KunenaCategoryHelper;
+use Kunena\Forum\Libraries\Forum\Message\KunenaMessageHelper;
 use Kunena\Forum\Libraries\Route\KunenaRoute;
 use Kunena\Forum\Libraries\User\KunenaUser;
 use Kunena\Forum\Libraries\User\KunenaUserHelper;
@@ -258,5 +259,75 @@ class UserController extends FormController
 		}
 
 		$this->saveInternal('apply');
+	}
+
+	/**
+	 * Move Messages
+	 *
+	 * @return  void
+	 *
+	 * @throws  Exception
+	 * @throws  null
+	 * @since   Kunena 2.0
+	 */
+	public function moveMessages(): void
+	{
+		if (!Session::checkToken())
+		{
+			$this->app->enqueueMessage(Text::_('COM_KUNENA_ERROR_TOKEN'), 'error');
+			$this->setRedirect(KunenaRoute::_($this->baseurl, false));
+
+			return;
+		}
+
+		$catid = $this->app->input->getInt('catid');
+		$uids  = (array) $this->app->getUserState('kunena.usermove.userids');
+
+		$error = null;
+
+		if ($uids)
+		{
+			foreach ($uids as $id)
+			{
+				list($total, $messages) = KunenaMessageHelper::getLatestMessages(false, 0, 0, ['starttime' => '-1', 'user' => $id]);
+
+				foreach ($messages as $object)
+				{
+					$topic = $object->getTopic();
+
+					if (!$object->isAuthorised('move'))
+					{
+						$error = $object->getError();
+					}
+					else
+					{
+						$target = KunenaCategoryHelper::get($catid);
+
+						if (!$topic->move($target, false, false, '', false))
+						{
+							$error = $topic->getError();
+						}
+					}
+				}
+			}
+		}
+		else
+		{
+			$this->app->enqueueMessage(Text::_('COM_KUNENA_PROFILE_NO_USER'), 'error');
+			$this->setRedirect(KunenaRoute::_($this->baseurl, false));
+
+			return;
+		}
+
+		if ($error)
+		{
+			$this->app->enqueueMessage($error, 'error');
+		}
+		else
+		{
+			$this->app->enqueueMessage(Text::_('COM_KUNENA_A_USERMES_MOVED_DONE'), 'success');
+		}
+
+		$this->setRedirect(KunenaRoute::_($this->baseurl, false));
 	}
 }
