+++ src/admin/tmpl/user/move.php
@@ -23,7 +23,7 @@ use Kunena\Forum\Libraries\Route\KunenaRoute;
 		<div id="j-main-container" class="col-md-12" role="main">
 			<form action="<?php echo KunenaRoute::_('administrator/index.php?option=com_kunena') ?>" method="post"
 				  id="adminForm" name="adminForm">
-				<input type="hidden" name="view" value="users"/>
+				<input type="hidden" name="view" value="user"/>
 				<input type="hidden" name="task" value=""/>
 				<input type="hidden" name="boxchecked" value="1"/>
 				<?php echo HTMLHelper::_('form.token'); ?>
